+++ apps/backend/src/Application/MemberFaces/Common/FaceSearchVectorRequestDto.cs
@@ -2,7 +2,7 @@ namespace backend.Application.MemberFaces.Common;
 
 public class FaceSearchVectorRequestDto
 {
-    public List<double> Embedding { get; set; } = null!;
+    public List<float> Embedding { get; set; } = null!;
     public string? FamilyId { get; set; }
     public string? MemberId { get; set; }
     public int TopK { get; set; } = 5;
